Add OpenAPIRenderer
